[Training] Unifying Preprocess + Postprocessing logic for Train/Oneshot (#1212)
Order of reviews:
https://github.com/vllm-project/llm-compressor/pull/1206
https://github.com/vllm-project/llm-compressor/pull/1207
https://github.com/vllm-project/llm-compressor/pull/1209
https://github.com/vllm-project/llm-compressor/pull/1212  <-- Here
https://github.com/vllm-project/llm-compressor/pull/1214

SUMMARY:
* Move the preprocessing and postprocessing logic out of
`src/llmcompressor/transformers/finetune/text_generation.py` and into
`src/llmcompressor/entrypoints/utils.py`

TEST PLAN:
Pass tests
