+++ src/llmcompressor/transformers/finetune/text_generation.py
@@ -17,22 +17,12 @@
 # Adapted from https://github.com/huggingface/transformers
 # vllm-project: no copyright
 
-import os
 import warnings
 from pathlib import PosixPath
-from typing import Optional
 
 from compressed_tensors.utils.helpers import deprecated
 from loguru import logger
-from transformers import (
-    AutoConfig,
-    AutoModelForCausalLM,
-    AutoProcessor,
-    HfArgumentParser,
-    PreTrainedModel,
-    set_seed,
-)
-from transformers.utils.quantization_config import CompressedTensorsConfig
+from transformers import HfArgumentParser
 
 from llmcompressor.args import (
     DatasetArguments,
@@ -41,11 +31,7 @@ from llmcompressor.args import (
     TrainingArguments,
 )
 from llmcompressor.core import reset_session
-from llmcompressor.pytorch.model_load.helpers import (
-    fallback_to_cpu,
-    parse_dtype,
-    save_checkpoint,
-)
+from llmcompressor.pytorch.model_load.helpers import save_checkpoint
 from llmcompressor.recipe import Recipe, StageRunType
 from llmcompressor.transformers.finetune.runner import StageRunner
 from llmcompressor.transformers.finetune.trainer import Trainer
@@ -53,14 +39,6 @@ from llmcompressor.transformers.sparsification.compressed_tensors_utils import (
     modify_save_pretrained,
     patch_tied_tensors_bug,
 )
-from llmcompressor.transformers.sparsification.sparse_model import (
-    get_processor_name_from_model,
-)
-from llmcompressor.transformers.utils.helpers import (
-    detect_last_checkpoint,
-    is_model_ct_quantized_from_path,
-)
-from llmcompressor.typing import Processor
 from llmcompressor.utils.fsdp.helpers import is_fsdp_model
 
 
@@ -70,15 +48,6 @@ def train(**kwargs):
     """
     model_args, dataset_args, recipe_args, training_args = parse_args(**kwargs)
     training_args.do_train = True
-    main(model_args, dataset_args, recipe_args, training_args)
-
-
-def eval(**kwargs):
-    """
-    CLI entrypoint for running evaluation
-    """
-    model_args, dataset_args, recipe_args, training_args = parse_args(**kwargs)
-    training_args.do_eval = True
     main(model_args, dataset_args, recipe_args, training_args)
 
 
@@ -98,10 +67,14 @@ def apply(**kwargs):
     """
     CLI entrypoint for any of training, oneshot
     """
-    report_to = kwargs.get("report_to", None)
-    model_args, dataset_args, recipe_args, training_args = parse_args(**kwargs)
+    from llmcompressor.args import parse_args
+
+    model_args, dataset_args, recipe_args, training_args, _ = parse_args(
+        include_training_args=True, **kwargs
+    )
 
     training_args.run_stages = True
+    report_to = kwargs.get("report_to", None)
     if report_to is None:  # user didn't specify any reporters
         # get rid of the reporters inferred from hugging face
         training_args.report_to = []
@@ -123,7 +96,6 @@ def parse_args(**kwargs):
             src/llmcompressor/transformers/utils/arg_parser/recipe_args.py
         * training_args in
             src/llmcompressor/transformers/utils/arg_parser/training_args.py
-
     """
     parser = HfArgumentParser(
         (ModelArguments, DatasetArguments, RecipeArguments, TrainingArguments)
@@ -161,147 +133,6 @@ def parse_args(**kwargs):
     return model_args, dataset_args, recipe_args, training_args
 
 
-def initialize_model_from_path(
-    model_args: ModelArguments,
-    training_args: Optional[TrainingArguments] = None,
-):
-    # Load pretrained model
-    # The .from_pretrained methods guarantee that only one local process can
-    # concurrently download model & vocab.
-    model_path = model_args.model
-    config = AutoConfig.from_pretrained(
-        model_args.config_name if model_args.config_name else model_path,
-        cache_dir=model_args.cache_dir,
-        revision=model_args.model_revision,
-        use_auth_token=True if model_args.use_auth_token else None,
-        tie_word_embeddings=model_args.tie_word_embeddings,
-        trust_remote_code=model_args.trust_remote_code_model,
-    )
-
-    last_checkpoint = None
-    teacher = None
-
-    if training_args is not None:
-        # Load teacher configuration if applicable
-        teacher_config = (
-            AutoConfig.from_pretrained(
-                model_args.distill_teacher,
-                use_auth_token=True if model_args.use_auth_token else None,
-                tie_word_embeddings=model_args.tie_word_embeddings,
-                trust_remote_code=model_args.trust_remote_code_model,
-            )
-            if model_args.distill_teacher
-            else None
-        )
-
-        # Detect last checkpoint
-        last_checkpoint = detect_last_checkpoint(training_args, model_args=model_args)
-
-        # Set seed before initializing model
-        set_seed(training_args.seed)
-
-        # Initialize teacher model if teacher path is provided
-        if model_args.distill_teacher is not None:
-            teacher_device_map = (
-                None
-                if os.environ.get("ACCELERATE_USE_FSDP", "false") == "true"
-                else "auto"
-            )
-            teacher_kwargs = {
-                "config": teacher_config,
-                "cache_dir": model_args.cache_dir,
-                "use_auth_token": True if model_args.use_auth_token else None,
-                "torch_dtype": parse_dtype(model_args.precision),
-                "device_map": teacher_device_map,
-                "trust_remote_code": model_args.trust_remote_code_model,
-            }
-
-            teacher = AutoModelForCausalLM.from_pretrained(
-                model_args.distill_teacher,
-                **teacher_kwargs,
-            )
-            if "sequence_length" in teacher_kwargs:
-                teacher.seqlen = teacher_kwargs["sequence_length"]
-
-    model_path = (
-        last_checkpoint or model_args.model
-        if hasattr(model_args, "model")
-        else model_args.model_name_or_path
-    )
-
-    # Fallback to CPU if GPU requested and not available
-    model_args.oneshot_device = fallback_to_cpu(model_args.oneshot_device)
-
-    # Trainer handles device assignment for FSDP and training, don't do mapping here
-    # if running oneshot outside of FSDP, apply user device settings
-
-    fsdp_enabled = os.environ.get("ACCELERATE_USE_FSDP", "false") == "true"
-
-    device_map = model_args.oneshot_device
-    if not fsdp_enabled and training_args is not None and training_args.do_train:
-        device_map = "auto"
-
-    model_kwargs = {
-        "config": config,
-        "cache_dir": model_args.cache_dir,
-        "revision": model_args.model_revision,
-        "use_auth_token": True if model_args.use_auth_token else None,
-        "torch_dtype": parse_dtype(model_args.precision),
-        "device_map": device_map,
-        "trust_remote_code": model_args.trust_remote_code_model,
-    }
-
-    # this calls from_pretrained under the hood so should be FSDP safe
-
-    # optimized models must be decompressed to carry out oneshot/train/etc
-    if is_model_ct_quantized_from_path(model_path):
-        model_kwargs["quantization_config"] = CompressedTensorsConfig(
-            run_compressed=False
-        )
-
-    model = AutoModelForCausalLM.from_pretrained(
-        model_path,
-        **model_kwargs,
-    )
-    if "sequence_length" in model_kwargs:
-        model.seqlen = model_kwargs["sequence_length"]
-
-    return model, teacher
-
-
-def initialize_processor_from_path(
-    model_args: ModelArguments,
-    model: PreTrainedModel,
-    teacher: Optional[PreTrainedModel] = None,
-) -> Processor:
-    processor_src = model_args.processor or get_processor_name_from_model(
-        model, teacher
-    )
-    # The use_fast=True option is not currently supported safely in Transformers
-    # See: https://github.com/huggingface/transformers/pull/34836#issuecomment-2491809727  # noqa: E501
-    try:
-        processor = AutoProcessor.from_pretrained(
-            processor_src,
-            cache_dir=model_args.cache_dir,
-            use_fast=True,
-            revision=model_args.model_revision,
-            use_auth_token=True if model_args.use_auth_token else None,
-            trust_remote_code=model_args.trust_remote_code_model,
-        )
-    except Exception:
-        logger.debug("Could not load fast processor, loading slow processor instead")
-        processor = AutoProcessor.from_pretrained(
-            processor_src,
-            cache_dir=model_args.cache_dir,
-            use_fast=False,
-            revision=model_args.model_revision,
-            use_auth_token=True if model_args.use_auth_token else None,
-            trust_remote_code=model_args.trust_remote_code_model,
-        )
-
-    return processor
-
-
 def main(
     model_args: ModelArguments,
     dataset_args: DatasetArguments,
@@ -326,10 +157,15 @@ def main(
 
     :param model_args: Arguments pertaining to which model/config/tokenizer we are
     going to fine-tune from
-    :param dataset_args: Arguments pertaining to what data we are going to input
-        our model for training
+    :param dataset_args: Arguments pertaining to what data we are
+        going to input our model for training
     :param training_args: Arguments pertaining to training loop configuration
     """
+    from llmcompressor.args import TrainingArguments
+    from llmcompressor.entrypoints.utils import (
+        initialize_model_from_path,
+        initialize_processor_from_path,
+    )
 
     # Temporary warning, to be removed
     if model_args.tie_word_embeddings is True:
@@ -426,6 +262,7 @@ def main(
 
         # exit immediately
         return
+
     # Training
     if training_args.do_train:
         checkpoint = None
